feat: scenario 2 tests
